Implement mouth map
Implementation of a function that calculates the mouth map for a given
image. Still experimenting on weather we should take into consideration
for the number of pixels in the face mask or not for our specific
purpose. Seems to work differently on different images. For now, we do
not use the face mask.